Handle `Transfer-Encoding`
Previously, this library would explicitly set the `Content-Length`
header on the passed-through response to the value of the
`ContentLength` property on the response object being passed.

However, `Content-Length` and `Transfer-Encoding` are mutually
exclusive - when `Transfer-Encoding` is used, the Go http package
will set the `ContentLength` property to `-1` to indicate no value.

This commit updates `go-passthrough` to handle this case. If the
`ContentLength` property is set to `-1`, the library will now
pass-through the `Transfer-Encoding` header instead of
`Content-Length`.

This commit is based on a fix developed by @tquinn86